% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-shinyWYSIWYG.R
\docType{methods}
\name{shinyWYSIWYG-GUI}
\alias{shinyWYSIWYG}
\alias{shinyWYSIWYG-GUI}
\title{shinyWYSIWYG GUI}
\usage{
shinyWYSIWYG()
}
\description{
shinyWYSIWYG package shiny-based graphical user interface (GUI)
\code{shinyWYSIWYG} starts the GUI.
}
\examples{
## Start the GUI
\dontrun{
shinyWYSIWYG();
}

}

